changes impleneted

+++ api/Core/DTOs/ItemMedia/UpdateItemMediaDto.cs
@@ -11,7 +11,7 @@ namespace Xcianify.Core.DTOs.ItemMedia
         public string MediaType { get; set; }
         public string FileName { get; set; }
         public string FileExtension { get; set; }
-        public int? FileSizeBytes { get; set; }
+        public long? FileSizeBytes { get; set; }
         public string MimeType { get; set; }
         public string MediaUrl { get; set; }
         public string Description { get; set; }
